requirements.txt
# Python 3.10.10 
absl-py==2.0.0
attrs==23.1.0
cffi==1.16.0
contourpy==1.2.0
cycler==0.12.1
djitellopy2==2.3
flatbuffers==23.5.26
fonttools==4.46.0
kiwisolver==1.4.5
matplotlib==3.8.2
mediapipe==0.10.9
numpy==1.26.2
opencv-contrib-python==4.8.1.78
opencv-python==4.8.1.78
packaging==23.2
Pillow==10.2.0
protobuf==3.20.3
pycparser==2.21
pyparsing==3.1.1
python-dateutil==2.8.2
six==1.16.0
sounddevice==0.4.6